# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateUserInput:
    properties:
      email:
        example: ivanov@example.com
        type: string
      full_name:
        example: Иванов Иван Иванович
        type: string
      password:
        example: securepass123
        type: string
      role:
        example: technician
        type: string
      username:
        example: ivanov
        type: string
    type: object
...
  handlers.UpdateUserInput:
    properties:
      email:
        example: new@example.com
        type: string
      full_name:
        example: Иванов Иван Иванович
        type: string
      is_active:
        example: true
        type: boolean
      password:
        example: newpassword
        type: string
      role:
        example: engineer
        type: string
    type: object
  handlers.UpdateWorkOrderInput:
    properties:
      status:
        example: in_progress
        type: string
    type: object
...
  models.Equipment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      installation_date:
        type: string
      inventory_number:
        type: string
      last_maintenance_date:
        type: string
      location:
        type: string
      manufacturer:
        type: string
      model:
        type: string
      name:
        type: string
      serial_number:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  models.RepairRequest:
    properties:
      assigned_to:
        type: integer
      assignee:
        $ref: '#/definitions/models.User'
      completed_at:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      creator:
        $ref: '#/definitions/models.User'
      description:
        type: string
      equipment:
        $ref: '#/definitions/models.Equipment'
      equipment_id:
        type: integer
      id:
        type: integer
      priority:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.WorkOrder:
    properties:
      actual_end:
        type: string
      actual_start:
        type: string
      assigned_to:
        type: integer
      assignee:
        $ref: '#/definitions/models.User'
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      planned_end:
        type: string
      planned_start:
        type: string
      repair_request:
        $ref: '#/definitions/models.RepairRequest'
      repair_request_id:
        type: integer
      schedule_id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      meta: {}
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API веб-приложения для управления техническим обслуживанием и ремонтом
    оборудования предприятия
  title: ТОиР API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Получение списка всех пользователей (только admin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Список пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя (только admin)
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Создание пользователя
      tags:
      - users
  /users/{id}:
    put:
      consumes:
      - application/json
      description: Обновляет данные пользователя (только admin)
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Обновление пользователя
      tags:
      - users
  /work-orders:
    get:
      description: Получение списка наряд-заказов с пагинацией и фильтрами
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Элементов на странице
        in: query
        name: per_page
        type: integer
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      - description: Фильтр по исполнителю
        in: query
        name: assigned_to
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.WorkOrder'
                  type: array
                meta:
                  $ref: '#/definitions/response.Meta'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Список наряд-заказов
      tags:
      - work-orders
    post:
      consumes:
      - application/json
      description: Создаёт новый наряд-заказ на выполнение работ
      parameters:
      - description: Данные наряд-заказа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.WorkOrder'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.WorkOrder'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Создание наряд-заказа
      tags:
      - work-orders
  /work-orders/{id}:
    put:
      consumes:
      - application/json
      description: Обновляет статус наряд-заказа
      parameters:
      - description: ID наряд-заказа
        in: path
        name: id
        required: true
        type: integer
      - description: Новый статус
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateWorkOrderInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.WorkOrder'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Обновление наряд-заказа
      tags:
      - work-orders
securityDefinitions:
  BearerAuth:
    description: 'Введите JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"